Add cfhCompile_Runtime_ClassLogger_Strategy_Database class.
The idea behind this is to allow a developer to log what classes are being declared by the application. In the future we can take this information an build a class iterator that reads the information from the database and then compiles a file containing the most widely used classes.